上游已把游戏改为 TF 卡加载；冲突按「架构收上游、脚位留实测」处理，并更新 journey 与 flash-roms 提示脚本。

+++ main/input_serial.h
@@ -10,5 +10,6 @@
 /* 装 UART0 驱动。必须在开始模拟之前调用一次。 */
 void input_serial_init(void);
 
-/* 每帧调一次，返回 NES 手柄位掩码（NES_PAD_* 的组合），直接喂给 input_update()。 */
-uint8_t input_serial_poll(void);
+/* 每帧调一次，返回宿主按键位掩码。低 8 位与 NES_PAD_* 一致，可直接喂给
+ * input_update()；高两位是 GAMEPAD_BIT_X/Y，只有 SNES 会用。 */
+uint16_t input_serial_poll(void);
